crop_recommendation/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flatbuffers==25.2.10
frozenlist==1.7.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
joblib==1.4.2
keras==3.10.0
libclang==18.1.1
lightgbm==4.6.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.4
namex==0.0.9
numpy==2.1.0
openai==0.28.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.2
protobuf==5.29.4
# psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pyngrok==7.2.5
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
# tensorboard==2.19.0
# tensorboard-data-server==0.7.2
# tensorflow==2.19.0
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
